Excel Sheet Example framework
This is the excel sheet I used to build the Apps Script for, so you should be able to see all the cell values and run a test using the code yourself. You'll find that Apps Script is actually only capable of running this code for 50 rows per time; I dealt with this by changing the cell ranges once the first 50 events had been created since this was faster than dealing with it in code.